requirements.txt

langchain==0.3.7
langchain-google-genai==2.0.4
langchain-community==0.3.7
langgraph==0.2.48
python-dotenv>=1.0.0
httpx>=0.24.1
duckduckgo-search==6.3.4

fastapi>=0.100.0,<0.104.0
pydantic==1.10.22
uvicorn==0.23.2



